moved msg_add_validator
