+++ x/stakeibc/keeper/msg_server.go
@@ -257,3 +257,20 @@ func (k msgServer) RegisterHostZone(goCtx context.Context, msg *types.MsgRegiste
 
 	return &types.MsgRegisterHostZoneResponse{}, nil
 }
+
+func (k msgServer) AddValidators(goCtx context.Context, msg *types.MsgAddValidators) (*types.MsgAddValidatorsResponse, error) {
+	ctx := sdk.UnwrapSDKContext(goCtx)
+
+	for _, validator := range msg.Validators {
+		if err := k.AddValidatorToHostZone(ctx, msg.HostZone, *validator, false); err != nil {
+			return nil, err
+		}
+
+		// Query and store the validator's sharesToTokens rate
+		if err := k.QueryValidatorSharesToTokensRate(ctx, msg.HostZone, validator.Address); err != nil {
+			return nil, err
+		}
+	}
+
+	return &types.MsgAddValidatorsResponse{}, nil
+}
