+++ x/stakeibc/keeper/msg_server_add_validators.go
@@ -1,26 +1,1 @@
 package keeper
-
-import (
-	"context"
-
-	sdk "github.com/cosmos/cosmos-sdk/types"
-
-	"github.com/Stride-Labs/stride/v18/x/stakeibc/types"
-)
-
-func (k msgServer) AddValidators(goCtx context.Context, msg *types.MsgAddValidators) (*types.MsgAddValidatorsResponse, error) {
-	ctx := sdk.UnwrapSDKContext(goCtx)
-
-	for _, validator := range msg.Validators {
-		if err := k.AddValidatorToHostZone(ctx, msg.HostZone, *validator, false); err != nil {
-			return nil, err
-		}
-
-		// Query and store the validator's sharesToTokens rate
-		if err := k.QueryValidatorSharesToTokensRate(ctx, msg.HostZone, validator.Address); err != nil {
-			return nil, err
-		}
-	}
-
-	return &types.MsgAddValidatorsResponse{}, nil
-}
